refactor(db-macros): move `GenProxyArgs` into expand module
